The sunburst was not adequately representing the internal time spent in a function. (That is, the time spent in a function but not in subfunctions.) To show this time I modified upload.py so that when a child with nodes is constructed a final leaf-child is added that represents the function's internal time.
To support this in the visualization I changed the node color to be based
solely on name and not care about the parent. That way the "child" representing
internal time has the same color as the parent. I also modified the mouseover
highlighting so that all nodes the the same filename/name are highlighted,
not just the one the mouse is over.