Normalize overly large children sets
There are times when children have calculated sizes larger than
their parent (because with the profiling information we have it is
hard to untangle just how much cumulative time of a program was spent
in a given tree). When this happens I don't do anything fancy,
only normalize the children's sizes so they exactly fit within the parent.