Move heirarchical JSON processing to a Worker
Building some call trees for the JSON worker can take a while and it
leaves the page unresponsive while the user could be browsing the
stats table or changing parameters. Moving that processing to a
Worker allows the page to stay responsive so that users can change
visualization parameters or browse the stats table.