Try to load JS libraries from CDN in worker
Making changes to the depth and cutoff selects in the app
tries to restart the web worker. This can fail if the
SnakeViz server has been shut down, as happens when
using the %snakeviz magic from IPython. The worker tries
to load Immutable and lodash from the snakeviz server and
errs because the server is down.

By loading from a CDN the worker can be restarted even if the
SnakeViz server has been turned off, assuming that the user is online
or has the libraries cached. The app will fall back on trying to
load the libraries from the SnakeViz server if the user is offline.

Users who are both offline and have the SnakeViz server off
can still get in trouble. To really solve this I'll need to
make the worker not rely on any outside libraries.