Allow the root node to be changed
Clicking on a row in the table will make that row become the new root.
This is useful if the default heuristic doesn't find the root you want.
Amongst other reasons, this can happen with multithreaded programs where
the root you really want is a thread that doesn't take the most CPU
time. It could also be handy if you can't immediately find the function
you're interested in on the plot.

I'm not a Javascript expert, so it's quite possible that there are
better ways to do this. This might be too prone to being accidentally
triggered, and possibly the Reset button should go back to the original
root instead of the current root.