Add "Reset Root" button and rename "Reset" to "Reset Zoom"
Reset Root is for when you want to go back to the default chosen root
function for the visualization. (This root can be changed by clicking
on rows in the data table.) Reset Zoom is for zooming back out
so that the current root function is at the center of the visualization.