Explain Python version requirements on error
SnakeViz must be installed and run via the same version major
version (e.g. 2 or 3) of Python as was used to generate the
profile being displayed. This isn't always clear to people,
so this adds a note explaining this when there's an error
loading profile data via the pstats.Stats class.

Fixes #103
Fixes #78