Adds support for Jupyter instances running remotely, also adds new parameter for visualizing existing profiling result.
The tornado server was not reachable when user ran Jupyter remotely (e.g. on a server with SSH tunnelling). Now the HTML is retrieved from tornado on the server side (hence requests lib dependency) and loaded into iframe.

Also added support for reusing an already existing profiling results with the -p parameter in the magics.